requirements.txt
django-govuk-template[forms,scss]==0.7
faker>=0.8
